Move Effect to own file, add parameters to Combine Effect
