+++ audioled/effects.py
@@ -13,103 +13,12 @@ from scipy.ndimage.filters import gaussian_filter1d
 from scipy.signal import lfilter
 
 import audioled.dsp as dsp
+import audioled.colors as colors
 import audioled.filtergraph as filtergraph
+from audioled.effect import Effect
 
 SHORT_NORMALIZE = 1.0 / 32768.0
 
-class Effect(object):
-    """
-    Base class for effects
-
-    Effects have a number of input channels and a number of output channels.
-    Before each processing the effect is updated.
-
-    Input values can be accessed by self._inputBuffer[channelNumber], output values
-    are to be written into self_outputBuffer[channelNumber].
-    """
-    def __init__(self):
-        self.__initstate__()
-
-    def __initstate__(self):
-        self._t = 0.0
-        try:
-            self._inputBuffer
-        except AttributeError:
-            self._inputBuffer = None
-        try:
-            self._outputBuffer
-        except AttributeError:
-            self._outputBuffer = None
-
-    def numOutputChannels(self):
-        """
-        Returns the number of output channels for this effect
-        """
-        raise NotImplementedError('numOutputChannels() was not implemented')
-
-    def numInputChannels(self):
-        """
-        Returns the number of input channels for this effect.
-        """
-        raise NotImplementedError('numInputChannels() was not implemented')
-
-    def setOutputBuffer(self,buffer):
-        """
-        Set output buffer where processed data is to be written
-        """
-        self._outputBuffer = buffer
-
-    def setInputBuffer(self, buffer):
-        """
-        Set input buffer for incoming data
-        """
-        self._inputBuffer = buffer
-
-    def process(self):
-        """
-        The main processing function:
-        - Read input data from self._inputBuffer
-        - Process data
-        - Write output data to self._outputBuffer
-        """
-        raise NotImplementedError('process() was not implemented')
-
-    async def update(self, dt):
-        """
-        Update timing, can be used to precalculate stuff that doesn't depend on input values
-        """
-        self._t += dt
-
-    def __cleanState__(self, stateDict):
-        """
-        Cleans given state dictionary from state objects beginning with __
-        """
-        for k in list(stateDict.keys()):
-            if k.startswith('_'):
-                stateDict.pop(k)
-        return stateDict
-
-    def __getstate__(self):
-        """
-        Default implementation of __getstate__ that deletes buffer, call __cleanState__ when overloading
-        """
-        state = self.__dict__.copy()
-        self.__cleanState__(state)
-        return state
-
-    def __setstate__(self, state):
-        self.__dict__.update(state)
-        self.__initstate__()
-
-    def updateParameter(self, stateDict):
-        self.__setstate__(stateDict)
-
-    def getParameter(self):
-        return {}
-
-    @staticmethod
-    def getParameterDefinition():
-        return {}
 
 
 class ShiftEffect(Effect):
@@ -222,7 +131,7 @@ class Append(Effect):
         self._outputBuffer[0] = state
 
 class Combine(Effect):
-    def __init__(self, mode='lightenOnly'):
+    def __init__(self, mode=colors.blend_mode_default):
         self.mode = mode
         self.__initstate__()
 
@@ -232,11 +141,37 @@ class Combine(Effect):
     def numOutputChannels(self):
         return 1
 
+    @staticmethod
+    def getParameterDefinition():
+        definition = {
+            "parameters": {
+                "mode": colors.blend_modes
+            }
+        }
+        return definition
+
+    def getParameter(self):
+        definition = self.getParameterDefinition()
+        definition['parameters']['mode'] = [self.mode] + [x for x in colors.blend_modes if x!=self.mode]
+        return definition
+
     def process(self):
         if self._inputBuffer is None or self._outputBuffer is None:
             return
-        if self.mode == 'lightenOnly':
-            self._outputBuffer[0] = np.maximum(self._inputBuffer[0], self._inputBuffer[1])
+        if not self._inputBufferValid(0) and not self._inputBufferValid(1):
+            # no input on any channels
+            self._outputBuffer[0] = None
+        elif self._inputBufferValid(0) and self._inputBufferValid(1):
+            # input on both channels
+            self._outputBuffer[0] = colors.blend(self._inputBuffer[0], self._inputBuffer[1], self.mode)
+        elif self._inputBufferValid(0):
+            # only channel 0 valid
+            self._outputBuffer[0] = self._inputBuffer[0]
+        elif self._inputBufferValid(1):
+            # only channel 1 valid
+            self._outputBuffer[0] = self._inputBuffer[0]
+        else:
+            self._outputBuffer[0] = None
 
 
 class AfterGlowEffect(Effect):
